requirements.txt
undetected-chromedriver
selenium
webdriver-manager
pandas
beautifulsoup4
lxml
schedule
openpyxl
requests
pymongo
python-dotenv

